move vault headings to the vault-v2 folder
