+++ apps/browser/src/vault/popup/components/vault-v2/vault-header/vault-header-v2.component.html
